add cyclemaplimit patch
details :
Prevents cyclemap vote spam.  A new server cvar sv_callvoteCyclemapWaitTime controls how many seconds must pass before another cyclemap vote is called after such a vote fails, regardless of which client is calling the votes.  By default sv_callvoteCyclemapWaitTime is 0, which results in the original behavior.  When a map cycles or reloads, the timer is reset and the server behaves as if a cyclemap vote hasn't been called yet.